refactor: centralize palette constants and drop dead exports
- Drop unused exports on HUE_SHIFT_MIN_DEG/HUE_SHIFT_RANGE_DEG in
  core/src/paletteUtils.ts (only pickDiversePalette uses them; no
  external importer)
- Move PALETTE_COUNT/HUE_SHIFT_MAX_DEG out of paletteAssigner.ts into
  server/src/constants.ts, matching the codebase convention that server
  magic numbers live in server/src/constants.ts
- Update paletteAssigner.ts and clientMessageHandler.ts to import the
  server constants; the palette helper no longer exports any constant
